Completed more problems, made example input file better

+++ Kattis/cpp/examples/input.cpp
@@ -1,10 +1,11 @@
 #include <algorithm>
 #include <iostream>
+#include <cmath>
 
 using namespace std;
 
 int main(void) {
-	long long a, b;
+	int a, b;
 
 	while (cin >> a >> b)
 		cout << abs(a - b) << endl;
